Add Node class, begin work on routing
Add a Node class interface to Plugin.  Begin working on Routing.
Next step is to figure out how to properly connect (creation of
buffers, calling lv2_port_connect_xxx) and then how to model the
connection using a Router class or ConnectionManager class.  We
need a better name for that class.