Make ports connectable and JACK support
This is a really naive approach at getting the ports to connect.  More
work needs to be put into the cross-jack connections.  Right now, all
the connections and JACK client are managed by main(), obviously these
need to be refactored. This is an important milestone, regardless of how
crappy the implementation may be.