Basic support for buffers
This allows for automatic aquisition of buffers for ports.  There are a
number of shortcomings with this technique.  The main thing to fix is to
make it so buffers are not reconnected if possible.  Then, maybe later,
we can make it more intelligent about reusing buffers (instead of
allocating a fresh one for every output port)

Also, this version doesn't perform much memory management.  Since the
system is in flux, it is hard to determine the proper management model
right now.