More improvements to buffer management
Instead of connecting every single period, this makes it so that the
connections will happen only once, when the the node is added to the
graph.

For JackPorts, the buffers of connected Ports are dynamically aquired
during process.  We still end up with a whole Buffer for each port, but
this can be fixed down the road with optimization to Compile and
BufferProvider.

Mixing is not yet supported

Next step is to fix the memory management and make sure the interfaces
still make sense
.