CompositeProcessor parents and Session root
Add parent() to Node to allow walking of parent nodes. Example, port ->
processor -> compositeProcessor -> compositeProcessor -> NULL.  This is
needed for the compiler to support connecting ports from between
CompositeProcessors.  However, the compiler extensions are not
implemented.

Additionally, make a CompositeProcessor the root of Session, so now
everything is nicely hierarchial.