Compiling nested-processors and wire control ports
Allow compiler to handle "outside dependencies", dependencies that are
external to the current CompositeProcessor.  Also, handle
nested-dependencies, or dependencies that are contained with the current
CompositeProcessor, but is nested several levels deep.  This might not
be the most efficient solution - but it works, we can profile later.

Also, ensure that connecting control-ports works.  We now have
controllers and automation!