Add aggregation library from Qt Creator
Aggregation library is handy for allowing components to respond to
changes caused by other components.  For example, when a component adds
an object to the extension manager's pool, all interested components can
use Aggregation::query<Foo> to see if the object is of interest instead
of using qobject_cast<Foo>.  The benefit is that an object (an
aggregation) can implement multiple interfaces without falling into the
multiple-inheritance QObject diamond of doom.