Begin migrating extension-system for Unison
The Qt Creator ExtensionSystem is not _exactly_ what we need or want.
First, we have to change literals that are hardcoded for Nokia or
QtCreator use.  We should also favor RDF over XML for the descriptors.
So, while we are at it, might as well update the filename style and
rename "PluginSpec" which is an awkward abbreviation to "PluginInfo" as
"Info" is used rather frequently throughout Qt itself.

Also, include Qt Creator's license in the directory, but it may be that
we want to remove the option of the commercial license at this point?