Integrate ExtensionSystem and start modularizing
Integrate the Aggregation and ExtensionSystem features from QtCreator.
Rename ExtensionSystem's "Plugins" to "Extensions" to ease confusion
between Unison Plugins and Audio Plugins such as LV2 and VST stuff.
More of the code will be changed as we adopt it.

Ideally most things will be implemented by means of Extensions.  As new
features are added, we will need to add the proper abstractions to Core.
The extension system's addObject function combined with the Aggregations
makes it really easy for plugins to add implementations to the global
pool of objects, which can be queried when it is time to do operations
supported by plugins.  This should become more obvious as Unison
matures, for now, look at some of Qt Creator's plugins.

The libs directory is for supporting code.  Our utility classes,
implementations of different file formats, sample processing,
oscillators, definitions for MIDI and such will live here. 3rd-Party
code will also live in the libs directory.