More modularizing, fix RPATHS and namespaces
Even more modularizing.  Now it is more apparent what belongs in
libunison and what belongs in the Core extension. Move things
appropriately.  Update namespaces on all modules.  Add an 'Internal'
namespace in each module for classes we do not want to expose.
Eventually, we should use the pImpl idiom in exported classes.

Also, had to play with RPATHs again.  This time, make it so Extensions
can link to the extensions they depend on.