Rearranging to allow for buffer-size changes
We need to be able to support changes in the buffer-size (and sampling
rate) at run-time.  This begins the changes needed for this.
Additionally, this commit moves the root-processor to the backend.  The
backend may not be the most ideal place (perhaps Engine is better?) but
placing it in the Backend is simple, and it should be easy enough to
change down the road.