RT-safe connections in the Patch
This is a quick demo to demonstrate Commands. This is a Connect Command.
It allows us to add connections to a patch in realtime. The patch is
recompiled and then swapped out at runtime.

The connections themselves, should probably be moved to Patch. but it
works in Port just fine, so we will just deal with this for now.  Also,
we now have an obvious need for internal vs non-internal APIs for
libunison, so, we will probably end up adding some D-Ptrs to hide the
internal API once everything is folded back into master.