Rework BufferProvider and Processor
Now, Processor is allowed to hold an instance to BufferProvider, assigned
from activate().  This makes it so ports have a reference to
bufferProvider that can be used during PortConnect/Disconnect.  This is
probably not the ideal solution, but it works for now.  The best
solution will be more obvious once we add support for buffer-reuse and
mixing for 1:N connections.

There are some obvious memory leaks.  To fix these, we need to be able
to throw some objects in a garbage-collector which will not run within
the processing thread.