Initial commit of (a broken) LADSPA extension
This is a skeleton for the LADSPA extension.  Most of the SLv2 code has
been maintained just to make it clear what happens in Lv2.  You are
of course free to trash all of that and start fresh.  Note, implement
the LADSPA plugin in the most obvious way. Essenetially, anything that
is Lv2-specific, you should do that in the way most appropriate for
LADSPA. But, stuff specific to unison. Like, managing buffers,
registering the buffer-provider. Indexing available plugins on
startup, should be preserved.

If you see something like "sladspa" ignore it, that is from sed. And
you may see some lv2 around.