Move ladspa.h to libs directory
Extensions other than the Ladspa extension may need ladspa.h.  For
example, an optional Ladspa-browser extension. Or maybe a plugin that
natively uses some ladpsa plugins internally for some reason.  At least,
this moves the 3rd party code (even if it is only an interface) out of
the extensions directory.