Single-threaded integration of new scheduler
This is a basic integration. A single worker runs in the Jack thread.
We need to add worker thread(s) and then have a waiting semaphore on
worker-threads, and a counting semaphore for the completion.

This commit has some hacks to make members visible where we should use
better OOP methods.  But this is a transitional version. Also, old
Schedules are currently leaked due to the lack of Command GC.