Begin a QIngen library for interfacing with Ingen
After a long debate, I am no longer going to bother writing the DSP
engine from scratch.  It was fun, it was rewarding, but there is much
work left to do.  Ingen is fairly mature and is extensible.  Dave spent
time rearranging the source tree for me, so now it is ready to use.

QIngen aims to be a Qt-based client for Ingen.  The normal Ingen client
is not ideal for at least one reason:  it doesn't use Qt's signal
system.  QIngen provides a client interface and model store using
QObjects.

Right now this is a tool to use _with_ Ingen.  It will probably evolve
into a complete wrapper and Ingen can be removed as a direct dependency
to unisonstudio.